# Improve performance of point equality checks