--- a/ecc/bls12-381/bandersnatch/point.go
+++ b/ecc/bls12-381/bandersnatch/point.go
@@ -296,10 +296,17 @@ func (p *PointProj) Equal(p1 *PointProj) bool {
 	if p.Z.IsZero() || p1.Z.IsZero() {
 		return false
 	}
-	var pAffine, p1Affine PointAffine
-	pAffine.FromProj(p)
-	p1Affine.FromProj(p1)
-	return pAffine.Equal(&p1Affine)
+
+	var lhs, rhs fr.Element
+	lhs.Mul(&p.X, &p1.Z)
+	rhs.Mul(&p1.X, &p.Z)
+	if !lhs.Equal(&rhs) {
+		return false
+	}
+	lhs.Mul(&p.Y, &p1.Z)
+	rhs.Mul(&p1.Y, &p.Z)
+
+	return lhs.Equal(&rhs)
 }
 
 // IsZero returns true if p=0 false otherwise